コマンドの切り出し
